merged runtime branch to main

+++ aide/agent.py
@@ -6,9 +6,8 @@ from typing import Any, Callable, List
 import humanize
 from aide.function import SearchArxiv, SearchPapersWithCode
 from aide.actions import Debug, Draft, Improve, Finish, SubmitReview
-from aide.utils.util import install_missing_libraries
 from .backend import query
-from .interpreter import ExecutionResult
+from .utils.execution_result import ExecutionResult
 from .journal import Journal, Node
 from .utils import data_preview
 from .utils.config import Config
@@ -431,8 +430,12 @@ class Agent:
     # For backward compatibility, need to change once the pipeline is verified
     async def step(self, exec_callback: ExecCallbackType = None, callback_manager=None):
         # clear the submission dir from previous steps
-        shutil.rmtree(self.cfg.workspace_dir / "submission", ignore_errors=True)
-        (self.cfg.workspace_dir / "submission").mkdir(exist_ok=True)
+
+        if not self.cfg.exec.use_modal:
+            shutil.rmtree(self.cfg.workspace_dir / "submission", ignore_errors=True)
+            (self.cfg.workspace_dir / "submission").mkdir(exist_ok=True)
+        else:
+            await callback_manager.execute_callback("remove_submission_directory")
 
         if not self.journal.nodes or self.data_preview is None:
             self.update_data_preview()
@@ -465,21 +468,33 @@ class Agent:
                 "stage_end", "Improve", supress_errors=True
             )
         if exec_callback:
-            exec_result = exec_callback(result_node.code, True)
+            exec_result = exec_callback(result_node.code)
         else:
             exec_result = await callback_manager.execute_callback(
-                "exec", result_node.code, True
+                "exec", result_node.code
             )
-
-        result_node = self.parse_exec_result(
+        result_node = await self.parse_exec_result(
             node=result_node,
             exec_result=exec_result,
             exec_callback=exec_callback,
             callback_manager=callback_manager,
+            use_modal=self.cfg.exec.use_modal,
         )
+        # TODO: Fix this to check submission when using modal. Also verify the cache_best_node function
         # handle final cases where we missed buggy nodes somehow
         if not result_node.is_buggy:
-            if not (self.cfg.workspace_dir / "submission" / "submission.csv").exists():
+            submission_exists = False
+            if self.cfg.exec.use_modal:
+                submission_exists = await callback_manager.execute_callback(
+                    "has_submission"
+                )
+            elif not (
+                self.cfg.workspace_dir / "submission" / "submission.csv"
+            ).exists():
+                submission_exists = False
+            else:
+                submission_exists = True
+            if not submission_exists:
                 result_node.is_buggy = True
                 result_node.metric = WorstMetricValue()
                 logger.info(
@@ -493,21 +508,7 @@ class Agent:
         if best_node is not None:
             if best_node.id == result_node.id:
                 logger.info(f"Node {result_node.id} is the best node so far")
-                best_solution_dir = self.cfg.workspace_dir / "best_solution"
-                best_solution_dir.mkdir(exist_ok=True, parents=True)
-                # copy submission/submission.csv to best_submission/submission.csv
-                best_submission_dir = self.cfg.workspace_dir / "best_submission"
-                best_submission_dir.mkdir(exist_ok=True, parents=True)
-                shutil.copy(
-                    self.cfg.workspace_dir / "submission" / "submission.csv",
-                    best_submission_dir,
-                )
-                # copy solution.py and relevant node id to best_solution/
-                with open(best_solution_dir / "solution.py", "w") as f:
-                    f.write(result_node.code)
-                # take note of the node id of the best node
-                with open(best_solution_dir / "node_id.txt", "w") as f:
-                    f.write(str(result_node.id))
+                await callback_manager.execute_callback("cache_best_node", result_node)
             else:
                 logger.info(f"Node {result_node.id} is not the best node")
                 logger.info(f"Node {best_node.id} is still the best node")
@@ -521,9 +522,9 @@ class Agent:
         max_attempts=3,
         exec_callback: ExecCallbackType = None,
         callback_manager=None,
+        use_modal=False,
     ) -> Node:
         logger.info(f"Agent is parsing execution results for node {node.id}")
-
         node.absorb_exec_result(exec_result)
 
         introduction = (
@@ -555,7 +556,6 @@ class Agent:
         if not isinstance(response, SubmitReview):
             logger.error(f"Expected SubmitReview but got {type(response)}")
             return None
-
         if (
             response.missing_libraries is not None
             and len(response.missing_libraries) > 0
@@ -565,13 +565,20 @@ class Agent:
                     f"Agent is missing libraries, attempting to install them: {response.missing_libraries}"
                 )
                 # install missing libraries
-                install_missing_libraries(response.missing_libraries)
+                await callback_manager.execute_callback(
+                    "install_dependecies", response.missing_libraries
+                )
+
                 # Re-execute the code after installing libraries
                 if exec_callback:
-                    exec_result = exec_callback(node.code, True)
+                    exec_result = exec_callback(
+                        node.code, True if not self.cfg.exec.use_modal else False
+                    )
                 else:
                     exec_result = await callback_manager.execute_callback(
-                        "exec", node.code, True
+                        "exec",
+                        node.code,
+                        True if not self.cfg.exec.use_modal else False,
                     )
                 # Recursively parse the new execution result
                 return await self.parse_exec_result(
@@ -581,24 +588,25 @@ class Agent:
                     callback_manager=callback_manager,
                     attempts=attempts + 1,
                     max_attempts=max_attempts,
+                    use_modal=use_modal,
                 )
             else:
                 logger.info(
                     "Maximum attempts reached while trying to install missing libraries"
                 )
-        else:
-            logger.info(
-                "Maximun attempts reached while trying to install missing libraries"
-            )
 
         # if the metric isn't a float then fill the metric with the worst metric
         if not isinstance(response.metric, float):
             response.metric = None
-
         # do an extra check, to catch cases where judge fails
-        has_csv_submission = (
-            self.cfg.workspace_dir / "submission" / "submission.csv"
-        ).exists()
+        if use_modal:
+            has_csv_submission = await callback_manager.execute_callback(
+                "has_submission"
+            )
+        else:
+            has_csv_submission = (
+                self.cfg.workspace_dir / "submission" / "submission.csv"
+            ).exists()
 
         node.analysis = response.summary
         node.is_buggy = (
